+++ Monitor/Monitor/Models/SerialMessages/WeatherData.cs
@@ -5,7 +5,7 @@ namespace Monitor.Models.SerialMessages;
 public class WeatherData : Message
 {
     [JsonPropertyName("temperature")]
-    public double Temperature { get; set; }
+    public float Temperature { get; set; }
 
     [JsonPropertyName("humidity")]
     public int Humidity { get; set; }
